+++ source/Core/IO/Stream.cpp
@@ -22,11 +22,11 @@ namespace core::io
         auto size = GetSize();
         if (m_cachedData.IsInvalid())
         {
-            m_cachedData.New(Alloc(size));
+            m_cachedData.New(Alloc(size_t(size)));
             Seek(0, SeekWhence::kSeekBegin);
             Read(m_cachedData->m_ptr, size);
         }
-        return { reinterpret_cast<const uint8_t*>(m_cachedData->m_ptr), size };
+        return { reinterpret_cast<const uint8_t*>(m_cachedData->m_ptr), uint32_t(size) };
     }
 
     SmartPtr<Stream> Stream::Next(bool isForced)
